Succesfully moved client v1 over to main rails app.
This involved adding skip_before_action :verify_authenticity_token to the API controller. This removed the requirement for A CSRF token on the POST request. CSRF tokens make sure that the post request is coming from your app. In order to use the API for more than just my web app, it makes more sense for me to authenticate API requests with Auth0 instead.